Move Eventbrite button logic into Eventbrite component
