--- common/styles/components/_event_schedule.scss
@@ -1,17 +1,0 @@
-.event-schedule__partnership {
-  @include respond-to('medium') {
-    text-align: right;
-  }
-}
-
-.event-schedule__free {
-  @include respond-to('large') {
-    text-align: right;
-  }
-}
-
-.event-schedule__main {
-  @include respond-to('large') {
-    border-right: 1px solid color('smoke');
-  }
-}
