+++ common/styles/non-critical.scss
@@ -9,7 +9,6 @@ $is-next: false !default;
 @import 'components/book_promo';
 @import 'components/control';
 @import 'components/divider';
-@import 'components/event_schedule';
 @import 'components/framed_header';
 @import 'components/feature_flags';
 @import 'components/find_us';
